move EngineMutex to stream_engine
